refactor validate testcase
